// fronts/students/src/app/models/school-contract.ts
import {RolContract} from "./users-model";

export interface SystemModuleContract {
    id: number;
    tag: string;
    module_name: string;
}
export interface SchoolModuleContract {
    id: number;
    school_id: number;
    system_module_id: number;
    status: boolean;
    is_active: boolean;
    system_module: SystemModuleContract;
}
export interface Schools {
    id: number;
    user_id: number;
    school_id: number;
    state: number;
    school: SchoolContract;
    roles: RolContract[];
}
export interface SchoolContract {
    id: number;
    country_id: number;
    statecode: string;
    nameschool: string;
    database_name: string;
    folder_name: string;
    lockdate: string;
    state: number;
    active: boolean;
}